refactor: put generateEmailContent in separate files
